chg: Move USE_DEFAULT to utils
The USE_DEFAULT class started its beautiful life
in the authenticators file, as the project has grown
its responsiblities have increased. We've decided to
move USE_DEFAULT into a new home under the utils
folder.

We hope USE_DEFAULT enjoys its new home with added
responsibility.
