+++ flask_rebar/authenticators/__init__.py
@@ -1,3 +1,3 @@
 from flask_rebar.authenticators.base import Authenticator
-from flask_rebar.authenticators.base import USE_DEFAULT
+from flask_rebar.utils.defaults import USE_DEFAULT
 from flask_rebar.authenticators.header_api_key import HeaderApiKeyAuthenticator
