+++ flask_rebar/swagger_generation/swagger_generator_v3.py
@@ -9,10 +9,10 @@
 """
 from __future__ import unicode_literals
 
+from flask_rebar.utils.defaults import USE_DEFAULT
 from flask_rebar.swagger_generation import swagger_words as sw
 from flask_rebar.swagger_generation.swagger_generator import SwaggerGenerator
 from flask_rebar.swagger_generation.swagger_objects import Server
-from flask_rebar.authenticators import USE_DEFAULT
 from flask_rebar.swagger_generation.generator_utils import (
     format_path_for_swagger,
     verify_parameters_are_the_same,
